Rewrite DVD shim to use ES content functions via /dev/es
Rewrite src/dvd_nand_shim.c to access content archive files via
/dev/es functions (ES_OpenContent, ES_ReadContent, ES_SeekContent)
instead of direct NAND file functions.

Ensure 32-byte alignment for ioctlv vectors and parameters, add
assembly wrappers in trampoline.S with 64-byte stack frame linkage,
and add -fno-stack-protector to CFLAGS in Makefile and convert.py.

+++ src/dvd_nand_shim.c
@@ -190,7 +190,7 @@ static s32 ES_ReadContent(s32 cfd, void* buf, u32 len);
 
 /* Dedicated static aligned buffers for ES vector IPC arguments */
 static s32 s_AlignedCfd __attribute__((aligned(32)));
-static s64 s_AlignedSeekOffset __attribute__((aligned(32)));
+static s32 s_AlignedSeekOffset __attribute__((aligned(32)));
 static s32 s_AlignedSeekWhence __attribute__((aligned(32)));
 static u16 s_AlignedContentIndex __attribute__((aligned(32)));
